pyproject.toml
[tool.poetry]
name = "ariadne-github-bot"
version = "0.1.0"
description = ""
authors = ["GraiaX-Community"]

[tool.poetry.dependencies]
python = "^3.9"
graia-ariadne = {extras = ["full"], version = "^0.5.0", allow-prereleases = true}
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"

[tool.black]
line-length = 99 # override black's default line-length
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
default = true
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"